@@ -5,7 +5,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.6.6</version>
+		<version>3.1.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.bank</groupId>
@@ -14,8 +14,8 @@
 	<name>accounts-service</name>
 	<description>Accounts service for bank application</description>
 	<properties>
-		<java.version>1.8</java.version>
-		<spring-cloud.version>2021.0.4</spring-cloud.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2022.0.4</spring-cloud.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -33,6 +33,7 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-config</artifactId>
+			<version>3.1.4</version>
 		</dependency>
 		<dependency>
 			<groupId>com.h2database</groupId>
