@@ -1,6 +1,5 @@
 package com.bank.accountsservice.data;
 
-import javax.annotation.PostConstruct;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,7 +18,7 @@ public class DataPopulator {
 			this.accountsRepository = accountsRepository;
 		}
 
-		@PostConstruct
+		
 		public void insertPropertyData() {
 			accountsRepository.save(new Customer("1001", "Alice"), new Accounts("20932K3150XC",100.0));
 			accountsRepository.save(new Customer("1002", "Bob"), new Accounts("20832K1450",101.23));
